feat(tab-scroller): typescript (#536)
